+++ src/App.jsx
@@ -1,11 +1,17 @@
-import MainPage from './components/MainContainer'
-import './style.scss'
+import { Route, Routes } from 'react-router-dom';
+import './style.scss';
+import MainPage from './components/MainContainer';
+import SecondForm from './components/SecondForm';
 
 function App() {
 
   return (
     <div className="AppCover">
-      <MainPage />
+
+      <Routes>
+        <Route path='/' element={<MainPage /> } />
+        <Route path='/second-form' element={<SecondForm />} />
+      </Routes>
     </div>
   )
 }
